+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # unreleased
+* Add local payment webhook support
 * Add subscription charged unsuccessfully sample webhook to webhook testing gateway
 * Fix issue where environment was not automatically converted from a string to a symbol (#148, #164)
 * Fix issue where logger could not be set on gateway instance